app/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
pydantic>=2.0
mcp[cli]>=1.2
surrealdb>=1.0.0
composio>=0.11.6
